requirements.txt
pycryptodome==3.20.0
pillow==10.3.0
pillow-heif==0.16.0
python-dotenv==1.0.0